+++ docs/changelog.rst
@@ -11,6 +11,9 @@ Changelog
 
 
 * Support Python 3
+* Added `--no-reset-stats` option for controling if the statistics should be reset once 
+  the hatching is complete
+* Added charts, for RPS and average response time, to the web UI
 * Various bug fixes and improvements